+++ src/util/strings.js
@@ -1,0 +1,48 @@
+// Copyright 2015-2019 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
+// @flow
+
+/*
+ * @dev Check if input is in Ascii table.
+ */
+export function isAscii(data) {
+	return /^[\x00-\x7F]*$/.test(data);
+}
+
+/*
+ * @dev Take hex encoded binary and make it utf-8 readable
+ */
+export function hexToAscii(hexx) {
+	const hex = hexx.toString();
+	let str = '';
+	for (let i = 0; i < hex.length; i += 2) {
+		str += String.fromCharCode(parseInt(hex.substr(i, 2), 16));
+	}
+
+	return str;
+}
+
+/*
+ * @dev Take a long string and output the first and last 10 chars.
+ */
+
+export function shortString(original) {
+	return original
+		.substr(0, 20)
+		.concat('......')
+		.concat(original.substr(original.length - 20));
+}
